[ONNX] Move no runtime tests out of runtime test cases (#80078)
- Moved no_runtime test cases from runtime script to no_runtime
Pull Request resolved: https://github.com/pytorch/pytorch/pull/80078
Approved by: https://github.com/BowenBao
